fixed logcat spam

+++ app/src/main/java/com/aboe/trivilauncher/data/remote/dto/forecast/WeatherComp.kt
@@ -25,7 +25,7 @@ data class WeatherComp (
             temp = main.temp,
             feelsLike = main.feelsLike,
             humidity = main.humidity,
-            windSpeed = wind.speed * 3.6,
+            windSpeed = (wind.speed * 3.6).toInt(),
             clouds = clouds.all,
             visibility = visibility
         )
